+++ source/Internal/Framework/Module/State/ModuleStateServiceInterface.php
@@ -17,20 +17,4 @@ interface ModuleStateServiceInterface
      * @return bool
      */
     public function isActive(string $moduleId, int $shopId): bool;
-
-    /**
-     * @param string $moduleId
-     * @param int    $shopId
-     *
-     * @throws ModuleStateIsAlreadySetException
-     */
-    public function setActive(string $moduleId, int $shopId);
-
-    /**
-     * @param string $moduleId
-     * @param int    $shopId
-     *
-     * @throws ModuleStateIsAlreadySetException
-     */
-    public function setDeactivated(string $moduleId, int $shopId);
 }