+++ tests/Integration/Core/Module/ModuleTest.php
@@ -38,8 +38,6 @@ class ModuleTest extends UnitTestCase
 
         $this->container->get('oxid_esales.module.install.service.launched_shop_project_configuration_generator')
             ->generate();
-
-        Registry::getConfig()->saveShopConfVar('aarr', 'activeModules', []);
     }
 
     public function testIsActiveIfModuleIsActive()