+++ tests/Integration/Internal/Framework/Module/Command/ModuleCommandsTestCase.php
@@ -9,13 +9,10 @@ declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Tests\Integration\Internal\Framework\Module\Command;
 
-use OxidEsales\EshopCommunity\Internal\Framework\Config\Dao\ShopConfigurationSettingDaoInterface;
-use OxidEsales\EshopCommunity\Internal\Framework\Config\DataObject\ShopConfigurationSetting;
-use OxidEsales\EshopCommunity\Internal\Framework\Config\DataObject\ShopSettingType;
+use OxidEsales\EshopCommunity\Internal\Container\ContainerFactory;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Install\DataObject\OxidEshopPackage;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Install\Service\ModuleInstallerInterface;
 use OxidEsales\EshopCommunity\Tests\Integration\Internal\Framework\Console\ConsoleTrait;
-use OxidEsales\EshopCommunity\Tests\Integration\Internal\ContainerTrait;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Application;
 use Webmozart\PathUtil\Path;
@@ -26,12 +23,29 @@ use Symfony\Component\Console\Tester\CommandTester;
  */
 class ModuleCommandsTestCase extends TestCase
 {
-    use ContainerTrait;
     use ConsoleTrait;
 
     protected $modulesPath = __DIR__ . '/Fixtures/modules/';
 
     protected $moduleId = 'testmodule';
+
+    private $container;
+
+    protected function tearDown(): void
+    {
+        $this->cleanupTestData();
+        ContainerFactory::resetContainer();
+        parent::tearDown();
+    }
+
+    protected function get(string $serviceId)
+    {
+        if ($this->container === null) {
+            $this->container = ContainerFactory::getInstance()->getContainer();
+        }
+
+        return $this->container->get($serviceId);
+    }
 
     /**
      * @return Application
@@ -53,15 +67,6 @@ class ModuleCommandsTestCase extends TestCase
                     Path::join($this->modulesPath, $this->moduleId)
                 )
             );
-
-        $activeModules = new ShopConfigurationSetting();
-        $activeModules
-            ->setName(ShopConfigurationSetting::ACTIVE_MODULES)
-            ->setValue([])
-            ->setShopId(1)
-            ->setType(ShopSettingType::ASSOCIATIVE_ARRAY);
-
-        $this->get(ShopConfigurationSettingDaoInterface::class)->save($activeModules);
     }
 
     protected function installTestModule(): void