+++ tests/Integration/Internal/Framework/Module/State/ModuleStateServiceTest.php
@@ -9,11 +9,11 @@ declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Tests\Integration\Internal\Framework\Module\State;
 
+use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Dao\ModuleConfigurationDaoInterface;
+use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\State\ModuleStateServiceInterface;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\State\ModuleStateIsAlreadySetException;
-use OxidEsales\EshopCommunity\Internal\Transition\Utility\ContextInterface;
 use OxidEsales\EshopCommunity\Tests\Integration\Internal\ContainerTrait;
-use OxidEsales\EshopCommunity\Tests\Unit\Internal\ContextStub;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -23,71 +23,18 @@ class ModuleStateServiceTest extends TestCase
 {
     use ContainerTrait;
 
-    private $moduleStateService;
-
-    public function setup(): void
+    public function testIsActive(): void
     {
-        $this->moduleStateService = $this->get(ModuleStateServiceInterface::class);
+        $configuration = new ModuleConfiguration();
+        $configuration
+            ->setModuleSource('test')
+            ->setActivated(true)
+            ->setId('testModule');
 
-        /** @var ContextStub $contextStub */
-        $contextStub = $this->get(ContextInterface::class);
-        $contextStub->setAllShopIds([1,2]);
+        $this->get(ModuleConfigurationDaoInterface::class)->save($configuration, 1);
 
-        if ($this->moduleStateService->isActive('testModuleId', 1)) {
-            $this->moduleStateService->setDeactivated('testModuleId', 1);
-        }
+        $moduleStateService = $this->get(ModuleStateServiceInterface::class);
 
-        if ($this->moduleStateService->isActive('testModuleId', 2)) {
-            $this->moduleStateService->setDeactivated('testModuleId', 2);
-        }
-
-        parent::setUp();
-    }
-
-    public function testSetActive()
-    {
-        $this->assertFalse(
-            $this->moduleStateService->isActive('testModuleId', 1)
-        );
-        $this->assertFalse(
-            $this->moduleStateService->isActive('testModuleId', 2)
-        );
-
-        $this->moduleStateService->setActive('testModuleId', 1);
-        $this->moduleStateService->setActive('testModuleId', 2);
-
-        $this->assertTrue(
-            $this->moduleStateService->isActive('testModuleId', 1)
-        );
-        $this->assertTrue(
-            $this->moduleStateService->isActive('testModuleId', 2)
-        );
-    }
-
-    public function testSetActiveIfActiveStateIsAlreadySet()
-    {
-        $this->expectException(\OxidEsales\EshopCommunity\Internal\Framework\Module\State\ModuleStateIsAlreadySetException::class);
-        $this->moduleStateService->setActive('testModuleId', 1);
-        $this->expectException(ModuleStateIsAlreadySetException::class);
-        $this->moduleStateService->setActive('testModuleId', 1);
-    }
-
-    public function testSetDeactivated()
-    {
-        $this->moduleStateService->setActive('testModuleId', 1);
-
-        $this->moduleStateService->setDeactivated('testModuleId', 1);
-
-        $this->assertFalse(
-            $this->moduleStateService->isActive('testModuleId', 1)
-        );
-    }
-
-    public function testSetDeactivatedIfActiveStateIsNotSet()
-    {
-        $this->expectException(ModuleStateIsAlreadySetException::class);
-        $this->moduleStateService = $this->get(ModuleStateServiceInterface::class);
-
-        $this->moduleStateService->setDeactivated('testModuleId', 1);
+        $this->assertTrue($moduleStateService->isActive('testModule', 1));
     }
 }